+++ .devcontainer/dev_launcher
@@ -1,4 +1,4 @@
 #!/bin/bash
 
 # adapt to package name
-interrogation_room
+irs
