+++ .devcontainer/docker-compose.yml
@@ -11,8 +11,7 @@ services:
         # [Choice] Install Node.js
         INSTALL_NODE: "true"
         NODE_VERSION: "lts/*"
-        # Please adapt to package name:
-        PACKAGE_NAME: "interrogation_room"
+        PACKAGE_NAME: "irs"
         # On Linux, you may need to update USER_UID and USER_GID below if not your local UID is not 1000.
         USER_UID: 1000
         USER_GID: 1000
