+++ irs/adapters/outbound/kafka_producer.py
