+++ irs/core/exceptions.py
