+++ setup.cfg
@@ -15,8 +15,8 @@
 
 [metadata]
 # Please adapt to package name:
-name = interrogation_room
-version = attr: interrogation_room.__version__
+name = irs
+version = attr: irs.__version__
 description = Interrogation Room Service - a short description
 url = https://github.com/ghga-de/interrogation-room-service
 long_description = file: README.md
@@ -49,7 +49,7 @@ python_requires = >= 3.9
 [options.entry_points]
 # Please adapt to package name:
 console_scripts =
-    interrogation-room-service = interrogation_room.__main__:run
+    interrogation-room-service = irs.__main__:run
 
 [options.extras_require]
 dev =
